# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Distimer Swagger API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/group": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Get All Joined Groups",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/groupctrl.groupDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Create Group",
                "parameters": [
                    {
                        "description": "createGroupReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/groupctrl.createGroupReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/groupctrl.groupDTO"
                        }
                    }
                }
            }
        },
...
        "/group/join": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Join Group with Invite Code",
                "parameters": [
                    {
                        "description": "joinReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/groupctrl.joinReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/groupctrl.groupDTO"
                        }
                    }
                }
            }
        },
        "/group/member/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Get All Group Members",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/groupctrl.affiliationDTO"
                            }
                        }
                    }
                }
            }
        },
...
        "/studylog": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "StudyLog"
                ],
                "summary": "Get All My Study Logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "count",
                        "name": "count",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/studylogctrl.myStudyLogDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "StudyLog"
                ],
                "summary": "Create Study Log",
                "parameters": [
                    {
                        "description": "createStudyLogReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/studylogctrl.createStudyLogReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/studylogctrl.myStudyLogDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "groupctrl.affiliationDTO": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string"
                },
                "joined_at": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "role": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "groupctrl.createGroupReq": {
            "type": "object",
            "required": [
                "invite_policy",
                "name",
                "nickname",
                "reveal_policy"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "description between 0 and 100"
                },
                "invite_policy": {
                    "type": "integer",
                    "maximum": 2,
                    "minimum": 0
                },
                "name": {
                    "type": "string",
                    "example": "name between 3 and 30"
                },
                "nickname": {
                    "type": "string",
                    "example": "nickname between 1 and 20"
                },
                "nickname_policy": {
                    "type": "string",
                    "example": "nickname_policy between 0 and 50"
                },
                "reveal_policy": {
                    "type": "integer",
                    "maximum": 2,
                    "minimum": 0
                }
            }
        },
        "groupctrl.groupDTO": {
            "type": "object",
            "properties": {
                "create_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invite_policy": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nickname_policy": {
                    "type": "string"
                },
                "reveal_policy": {
                    "type": "integer"
                }
            }
        },
        "groupctrl.joinReq": {
            "type": "object",
            "required": [
                "invite_code"
            ],
            "properties": {
                "invite_code": {
                    "type": "string"
                }
            }
        },
...
        "studylogctrl.createStudyLogReq": {
            "type": "object",
            "required": [
                "content",
                "end_at",
                "groups_to_share",
                "start_at",
                "subject_id"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "example": "content between 0 and 30"
                },
                "end_at": {
                    "type": "string",
                    "example": "2020-08-28T09:20:26.187+09:00"
                },
                "groups_to_share": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "group_id"
                    ]
                },
                "start_at": {
                    "type": "string",
                    "example": "2020-08-28T09:20:26.187+09:00"
                },
                "subject_id": {
                    "type": "string",
                    "example": "subject_id"
                }
            }
        },
...
        "studylogctrl.myStudyLogDTO": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "end_at": {
                    "type": "string"
                },
                "groups_to_share": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "start_at": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}